feat: add code-section
close #22
